backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic[email]>=2.0.0
python-multipart>=0.0.6
alembic>=1.11.0
jinja2>=3.1.0
python-dotenv>=1.0.0
# For SQLite, no separate driver needed as it's built-in.
# If using PostgreSQL, you would add: psycopg2-binary
